+++ DicomTypeTranslation.Tests/DatabaseExamples.cs
@@ -37,7 +37,8 @@ public class DatabaseExamples : DatabaseTests
         tbl.AddColumn("FileLocation", new DatabaseTypeRequest(typeof(string), 500), true, 500);
 
         //Create a DataTable in memory for the data we read from disk
-        var dt = new DataTable();
+        using var dt = new DataTable();
+        dt.BeginLoadData();
         dt.Columns.Add("SOPInstanceUID");
         dt.Columns.Add("Modality");
         dt.Columns.Add("PatientID");
@@ -61,6 +62,7 @@ public class DatabaseExamples : DatabaseTests
 
         //put the DataTable into the database
         using var insert = tbl.BeginBulkInsert();
+        dt.EndLoadData();
         insert.Upload(dt);
 
 
@@ -85,10 +87,8 @@ public class DatabaseExamples : DatabaseTests
         //decide where you want to create the table
         var db = GetTestDatabase(dbType);
 
-        var creator = new ImagingTableCreation(db.Server.GetQuerySyntaxHelper());
-
         //actually do it
-        creator.CreateTable(db.ExpectTable("MyCoolTable"),toCreate);
+        ImagingTableCreation.CreateTable(db.ExpectTable("MyCoolTable"),toCreate);
     }
 
     public void TestGetDataTable()
@@ -101,7 +101,7 @@ public class DatabaseExamples : DatabaseTests
             new DicomDate(DicomTag.PatientBirthDate,new DateTime(2001,1,1))
         });
 
-        var dt = new DataTable();
+        using var dt = new DataTable();
         var row = ds.ToRow(dt);
 
         Assert.AreEqual("Frank", row["PatientName"]);